Add append_function API
Adds `append_function` that lets ScriptLanguage implementations decide
how a function is appended to a script, by default it inserts the
function at the end of the file (this was the previous behavior for all
languages).

Also implements C# append_function so the editor can properly add
functions when connecting signals.