Add `Object::has_property` method
Add `has_property` method to `Object` that checks if a property exists.
Same as the `in` keyword in GDScript but implemented as a method
registered in ClassDB so it can be used by other language bindings such
as GDExtension and C#.

The `{add,remove,has}_property` and `get_properties` have been renamed
to the more specific `{add,remove,has}_replication_property` and
`get_replication_properties` to avoid conflicts with the new `has_property`
method.