C#: Allow implicitly converting enums to Variant
The downside of this implementation is that enums will be boxed.
I think this is the reason why this wasn't implemented originally, but
it seems using enums with APIs that take Variant is a common issue and
it's difficult for users to figure out that they should just convert to
the underlying integral numeric type.