Implement new `dotnet` module
- Implements the `dotnet` module that will eventually replace the `mono` module.
- This module implements a `GDExtensionLoader` to load .NET assemblies as GDExtensions.
- The module implements loading .NET assemblies using hostfxr, monovm, and NativeAOT.
- In an editor build, .NET assemblies are loaded using the `Godot.PluginLoader` package
and hostfxr so they can be reloaded. The `Godot.EditorIntegration` GDExtension is
also loaded to enhance the user experience. The source of this packages is in the
`godot-dotnet` repo and they are fetched from a NuGet feed using the
`RestoreEditorPackages.proj` MSBuild project.
- The module exposes additional GDExtension API (`editor_internal`) needed by the
`Godot.EditorIntegration` GDExtension to provide all the features.