[.NET] Use compat system for virtual methods
Use the registered compatibility virtual methods to generate overloads in the C# bindings so we can attempt to preserve compatibility.

Since virtual methods are called using the `CSharpInstance::callp` method and we don't have a hash, we just check that the method name and arg count match to determine which overload to call.